# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ultra Venues API Docs",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/venues": {
      "get": {
        "tags": [
          "Venue"
        ],
        "summary": "List all venues",
        "description": "This endpoint retrieves a list of all venues, excluding sensitive information like password and isActive status.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/venues/{id}": {
      "get": {
        "tags": [
          "Venue"
        ],
        "summary": "Get venue details",
        "description": "This endpoint retrieves details of a specific venue by its ID, excluding sensitive information like password and isActive status.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the venue to retrieve details for"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Venue"
        ],
        "summary": "Update venue details",
        "description": "This endpoint updates the details of a specific venue by its ID, including uploading new pictures.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the venue to update"
          },
          {
            "name": "name",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "mobile": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "website": {
                  "example": "any"
                },
                "services": {
                  "example": "any"
                }
              }
            },
            "description": "Name of the venue"
          },
          {
            "name": "type",
            "in": "body",
            "description": "Type of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "email",
            "in": "body",
            "description": "Email of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "mobile",
            "in": "body",
            "description": "Mobile number of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "location",
            "in": "body",
            "description": "Location of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "website",
            "in": "body",
            "description": "Website URL of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "services",
            "in": "body",
            "description": "Services provided by the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin login endpoint",
        "description": "This endpoint allows an admin to login by providing email and password. On successful login, it returns access and refresh tokens.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Admin login credentials",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "admin@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/register": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin registration endpoint",
        "description": "This endpoint allows an admin to register by providing full name, email, phone number, and password. On successful registration, it returns a success message.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Admin registration details",
            "schema": {
              "type": "object",
              "properties": {
                "fullName": {
                  "type": "string",
                  "example": "John Doe"
                },
                "email": {
                  "type": "string",
                  "example": "admin@example.com"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "1234567890"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/refresh-token": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Refresh token endpoint",
        "description": "This endpoint allows an admin to refresh their access token using a refresh token.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Refresh token",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string",
                  "example": "some-refresh-token"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/customers": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all customers with booking count",
        "description": "This endpoint returns all customers along with their total booking count. Admin authentication is required.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/customers/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get customer details",
        "description": "This endpoint returns the details of a particular customer by their ID. Admin authentication is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/venues": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all venues",
        "description": "This endpoint returns all venues. Admin authentication is required.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/venues/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get venue details",
        "description": "This endpoint returns the details of a particular venue by its ID. Admin authentication is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a venue",
        "description": "This endpoint allows an admin to delete a venue by its ID. Admin authentication is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/venues/register/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Register a venue",
        "description": "This endpoint allows an admin to register a venue by setting its isActive status to true. Admin authentication is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}